Split functionality into different files per PR feedback
